Implement functions that checks the passed http status code value is in the expected range
This adds functions that checks the passed http status code value is in the expected range for `HttpStatus.***`
Names of each of functions are sorted with the spec. They uses `is<Semantics><NumberRange>` convention.

- [`HttpStatus.isInformational1xx`](https://httpwg.org/specs/rfc9110.html#status.1xx)
- [`HttpStatus.isSuccessful2xx`](https://httpwg.org/specs/rfc9110.html#status.2xx)
- [`HttpStatus.isRedirection3xx`](https://httpwg.org/specs/rfc9110.html#status.3xx)
- [`HttpStatus.isClientError4xx`](https://httpwg.org/specs/rfc9110.html#status.4xx)
- [`HttpStatus.isServerError5xx`](https://httpwg.org/specs/rfc9110.html#status.5xx)

We take an approach for unit test that is similar to
[property based testing](https://en.wikipedia.org/wiki/Software_testing#Property_testing).
The test case input values that satisfies a conditions (For http status code. It's very very simple condition).
And all test case files must check input patterns has no problem (no overlap, no missing spans) before running tests in them.