Update TypeScript 5.7
This change would not break the typecheck on user's CI since this release does not contain a change
which uses TypeScript's newer feature.

Probably user doesn't have to update your TypeScript to 5.7 or later _immediately_ before import this change.

However, we may begin to use all new features in TS 5.7.
**We recommend strongly to update TypeScript for your project.** We would not provide a support with TS5.6 or earlier.

- https://devblogs.microsoft.com/typescript/announcing-typescript-5-6/
- https://devblogs.microsoft.com/typescript/announcing-typescript-5-7/